changed int to unsigned long

+++ src/Bellman-Ford.cpp
@@ -19,11 +19,11 @@ using namespace std;
 
 
 //This void changes the paths from start to every other point, so that (if there exists a connection) in the end the shortest path from start to every connected point is shown.
-void newPaths(vector<vector<double>>& graph, int start){
-	int size=graph.size();
-	for (int next=0;next<size;next++){
+void newPaths(vector<vector<double>>& graph, unsigned long start){
+	unsigned long size=graph.size();
+	for (unsigned long next=0;next<size;next++){
 		if(graph[start][next]==graph[start][next]){    //Check if you need the path from i to other points. If start and i are not connected then ignore 
-			for (int end=0;end<size;end++){
+			for (unsigned long end=0;end<size;end++){
 				if(graph[next][end]==graph[next][end]){										                 //
 					if(graph[start][end]>graph[next][end]+graph[start][next] || graph[start][end]!=graph[start][end]){  //together in one if? Don't know how to say (maybe if(graph[i][j]!=NULL && (graph[start][j]<graph[i][j] || graph[start][j]==NULL)) ?)
 						if (graph[start][end]!=graph[start][end] && end<next) {
@@ -41,17 +41,17 @@ void newPaths(vector<vector<double>>& graph, int start){
 
 //Nothing special, just uses newPaths on the whole graph
 void newGraph(vector<vector<double>>& graph){
-	int size=graph.size();
-	for (int i=0; i<size;i++) newPaths(graph,i);
+	unsigned long size=graph.size();
+	for (unsigned long i=0; i<size;i++) newPaths(graph,i);
 }
 
 
 
 //if start has the value -INF, this function changes every successive point to -INF. This includes graph[start][start].
-void makeInf(vector<vector<double>>& graph,int start){
+void makeInf(vector<vector<double>>& graph,unsigned long start){
 	if (graph[start][start]!=-std::numeric_limits<double>::infinity()) return;
-	int size=graph.size();
-	for (int j=0;j<size;j++){
+	unsigned long size=graph.size();
+	for (unsigned long j=0;j<size;j++){
 		if (graph[start][j]==graph[start][j]){
 			graph[start][j]=-std::numeric_limits<double>::infinity();
 		} 
@@ -62,9 +62,9 @@ void makeInf(vector<vector<double>>& graph,int start){
 //In addition it changes (with makeInf() ) all related points to -INF)
 bool checkNeg(vector<vector<double>>& graph){
 	bool info=false;
-	int size=graph.size();
-	for (int i=0;i<size;i++){
-		for (int j=0;j<size;j++){
+	unsigned long size=graph.size();
+	for (unsigned long i=0;i<size;i++){
+		for (unsigned long j=0;j<size;j++){
 			if (graph[i][j]==graph[i][j] && graph[j][i]==graph[j][i]){
 				if(graph[i][j]+graph[j][i]<0 || graph[j][i]==-std::numeric_limits<double>::infinity() || graph[i][j]==-std::numeric_limits<double>::infinity()){
 					info=true;
@@ -79,9 +79,9 @@ bool checkNeg(vector<vector<double>>& graph){
 
 //simple Algorithm to show the matrix graph in a proper format
 void Output(vector<vector<double>> graph){ 
-	int size=graph.size();
-	for (int i=0;i<size;i++){
-		for (int k=0;k<size;k++){
+	unsigned long size=graph.size();
+	for (unsigned long i=0;i<size;i++){
+		for (unsigned long k=0;k<size;k++){
 			if (graph[i][k]!=graph[i][k]) printf("%5s ","N");
 			else if (graph[i][k]==-std::numeric_limits<double>::infinity()) printf ("%5s ", "-INF");
 			else	printf("%5.0f ",graph[i][k]);
@@ -95,9 +95,9 @@ void Output(vector<vector<double>> graph){
 //note that it's not necessary to check if one row of the original mtraix is empty, because the elements on the diagonal are never empty.
 CSR createCSR(vector<vector<double>> graph){
 	CSR matrix;
-	int old_i=0;
-	for (int i=0;i<graph.size();i++){
-		for (int k=0;k<graph.size();k++){
+	unsigned long old_i=0;
+	for (unsigned long i=0;i<graph.size();i++){
+		for (unsigned long k=0;k<graph.size();k++){
 			if (graph[i][k]==graph[i][k]) {
 				matrix.value.push_back(graph[i][k]);
 				matrix.row_idx.push_back(k);
@@ -115,16 +115,16 @@ CSR createCSR(vector<vector<double>> graph){
 //in this case we assume a square-matrix
 vector <vector<double>> createNormal(CSR matrix){
 	vector <vector<double>> normal;
-	int size=matrix.col_idx.size();
+	unsigned long size=matrix.col_idx.size();
 	normal.resize(size,vector<double>(size,std::numeric_limits<double>::quiet_NaN()));
-	for (int i=0;i<size;i++){
+	for (unsigned long i=0;i<size;i++){
 		if (i==size-1){ 
-			for (int j=matrix.col_idx[i];j<matrix.value.size();j++){
+			for (unsigned long j=matrix.col_idx[i];j<matrix.value.size();j++){
 				normal[i][matrix.row_idx[j]]=matrix.value[j];
 			}
 		}
 		else {
-			for (int j=matrix.col_idx[i];j<matrix.col_idx[i+1];j++){
+			for (unsigned long j=matrix.col_idx[i];j<matrix.col_idx[i+1];j++){
 				normal[i][matrix.row_idx[j]]=matrix.value[j];
 			}
 		}
@@ -134,24 +134,24 @@ vector <vector<double>> createNormal(CSR matrix){
 
 //generates Output of a CSR matrix in it's  own form
 void OutputCSR(CSR matrix){
-	for (int i=0;i<matrix.value.size();i++) printf("%5.0f",matrix.value[i]);
+	for (unsigned long i=0;i<matrix.value.size();i++) printf("%5.0f",matrix.value[i]);
 	printf("\n");
-	for (int i=0;i<matrix.row_idx.size();i++) printf("%5d",matrix.row_idx[i]);
+	for (unsigned long i=0;i<matrix.row_idx.size();i++) printf("%5d",matrix.row_idx[i]);
 	printf("\n");
-	for (int i=0;i<matrix.col_idx.size();i++) printf("%5d",matrix.col_idx[i]);
+	for (unsigned long i=0;i<matrix.col_idx.size();i++) printf("%5d",matrix.col_idx[i]);
 	printf("\n\n\n");
 
 }
 
 //checks a CSR-matrix if start reaches end directly, if yes changes idx to that index
 //if not changes idx to the index where this path belongs
-bool XreachesY(int start,int end,int& idx,CSR graph){
-	int size=graph.col_idx.size();
-	int ending = graph.row_idx.size();
+bool XreachesY(unsigned long start,unsigned long end,unsigned long& idx,CSR graph){
+	unsigned long size=graph.col_idx.size();
+	unsigned long ending = graph.row_idx.size();
 	if (start<size-1) ending=graph.col_idx[start+1];
 	idx=graph.col_idx[start];
-	for (int ptr=graph.col_idx[start]; ptr<ending; ptr++){
-		int test=graph.row_idx[ptr];
+	for (unsigned long ptr=graph.col_idx[start]; ptr<ending; ptr++){
+		unsigned long test=graph.row_idx[ptr];
 		if (test<end) idx=ptr;													//in case there's no connection: we set idx
 		if (graph.row_idx[ptr]==end){
 			idx=ptr;
@@ -162,23 +162,23 @@ bool XreachesY(int start,int end,int& idx,CSR graph){
 }
 
 //newPaths for CSR
-void newPathsCSR(CSR& graph, int start){
-	int size=graph.col_idx.size();
-	int ending1=graph.row_idx.size();
+void newPathsCSR(CSR& graph, unsigned long start){
+	unsigned long size=graph.col_idx.size();
+	unsigned long ending1=graph.row_idx.size();
 	if (start<size-1) ending1=graph.col_idx[start+1];
 
-	for (int next_point=graph.col_idx[start];next_point<ending1; next_point++){    //check every point, which is reachable from start
-		int ending2=graph.row_idx.size();
+	for (unsigned long next_point=graph.col_idx[start];next_point<ending1; next_point++){    //check every point, which is reachable from start
+		unsigned long ending2=graph.row_idx.size();
 		if (graph.row_idx[next_point]<size-1) ending2=graph.col_idx[graph.row_idx[next_point]+1];
 
-		for(int ptr = graph.col_idx[graph.row_idx[next_point]]; ptr<ending2; ptr++){    //check every point, which is reachable from next_point
-			int end=graph.row_idx[ptr];
-			int idx=0;
+		for(unsigned long ptr = graph.col_idx[graph.row_idx[next_point]]; ptr<ending2; ptr++){    //check every point, which is reachable from next_point
+			unsigned long end=graph.row_idx[ptr];
+			unsigned long idx=0;
 			double sum=graph.value[next_point] + graph.value[ptr];   //sum = path from start to next_point + path from next_point to end
 			if (XreachesY(start,end,idx,graph)){
 				if (graph.value[idx]>sum) graph.value[idx]=sum;
 			}else {
-				for (int i=start+1;i<size;i++) graph.col_idx[i]++;
+				for (unsigned long i=start+1;i<size;i++) graph.col_idx[i]++;
 				++ending2;
 				++ending1;
 				if(idx<next_point) ++next_point;	
@@ -196,22 +196,22 @@ void newPathsCSR(CSR& graph, int start){
 
 //newGraph for CSR
 void newGraphCSR(CSR& graph){
-	int size=graph.col_idx.size();
-	for (int i=0;i<size;i++) newPathsCSR(graph,i);
+	unsigned long size=graph.col_idx.size();
+	for (unsigned long i=0;i<size;i++) newPathsCSR(graph,i);
 }
 
 
 
 //newGraph for CSR, which, if too many paths exist, will change the CSR to a matrix.
 void newGraphCSR2(CSR& graph){
-	int size=graph.col_idx.size();
-	int critical_length = size / 10 ;     //at this point we will change the function to newGraph()
-	for (int i=0;i<size;i++) {
+	unsigned long size=graph.col_idx.size();
+	unsigned long critical_length = size / 10 ;     //at this point we will change the function to newGraph()
+	for (unsigned long i=0;i<size;i++) {
 		newPathsCSR(graph,i);
 		if ((graph.value.size()/ size) >= critical_length){
 			vector <vector<double>> matrix;
 			matrix = createNormal(graph);
-			for (int j=i;j<size;j++) newPaths(matrix,j);
+			for (unsigned long j=i;j<size;j++) newPaths(matrix,j);
 			graph=createCSR(matrix);
 			return;
 		}
@@ -222,12 +222,12 @@ void newGraphCSR2(CSR& graph){
 
 
 //if start has the value -INF, this function changes every successive point to -INF
-void makeInfCSR(CSR& graph,int start,int self_path){
+void makeInfCSR(CSR& graph,unsigned long start,unsigned long self_path){
 	if (graph.value[self_path]!=-std::numeric_limits<double>::infinity()) return;
-	int size=graph.col_idx.size();
-	int ending=graph.row_idx.size();
+	unsigned long size=graph.col_idx.size();
+	unsigned long ending=graph.row_idx.size();
 	if(start<size-1) ending=graph.col_idx[start+1];
-	for (int j=graph.col_idx[start];j<ending;j++){
+	for (unsigned long j=graph.col_idx[start];j<ending;j++){
 		graph.value[j]=-std::numeric_limits<double>::infinity();
 	}
 }
@@ -236,12 +236,12 @@ void makeInfCSR(CSR& graph,int start,int self_path){
 //In addition it changes (with makeInf() ) all related points to -INF)
 bool checkNegCSR(CSR& graph){
 	bool info=false;
-	int size=graph.col_idx.size();
-	for (int i=0;i<size;i++){
-		int ending=graph.row_idx.size();
+	unsigned long size=graph.col_idx.size();
+	for (unsigned long i=0;i<size;i++){
+		unsigned long ending=graph.row_idx.size();
 		if(i<size-1) ending=graph.col_idx[i+1];
-		for (int j=graph.col_idx[i];j<ending;j++){
-			int k=graph.col_idx[graph.row_idx[j]];
+		for (unsigned long j=graph.col_idx[i];j<ending;j++){
+			unsigned long k=graph.col_idx[graph.row_idx[j]];
 			while (graph.row_idx[k]<i && k<graph.row_idx.size()-1) k++; //checks if the path from j to i exists
 			if(graph.row_idx[k]==i){
 				if(graph.value[j]+graph.value[k]<0 || graph.value[j]==-std::numeric_limits<double>::infinity() || graph.value[k]==-std::numeric_limits<double>::infinity()){
@@ -274,10 +274,10 @@ void AllCSR2(CSR& graph){
 
 void AllOpt(vector <vector<double>>& graph){
 	CSR matrix=createCSR(graph);
-	int size = matrix.col_idx.size();
-	int critical_length= size / 10;   //further testing needed
+	unsigned long size = matrix.col_idx.size();
+	unsigned long critical_length= size / 10;   //further testing needed
 	bool test = true;
-	for (int i=0; i<size; i++){
+	for (unsigned long i=0; i<size; i++){
 		if(test){ 
 			newPathsCSR(matrix,i);
 			if ((matrix.value.size()/ size) >= critical_length){
